Adicionando login, logout, etc

+++ REST_API/resources/item.py
@@ -1,7 +1,7 @@
-import uuid
 from flask import request
 from flask.views import MethodView
 from flask_smorest import Blueprint, abort
+from flask_jwt_extended import jwt_required, get_jwt
 from sqlalchemy.exc import SQLAlchemyError
 
 from db import db
@@ -11,20 +11,29 @@ from schemas import ItemSchema, ItemUpdateSchema
 blp = Blueprint("Items", "items", description="Operations on items")
 
 
-@blp.route("/item/<string:item_id>")
+@blp.route("/item/<int:item_id>")
 class Item(MethodView):
+    # retorna um item
     @blp.response(200, ItemSchema)
     def get(self, item_id):
         item = ItemModel.query.get_or_404(item_id)
         return item
 
+    # deleta um item
+    @jwt_required()
     def delete(self, item_id):
+        # verifica se o usuário é admin
+        jwt = get_jwt()
+        if not jwt.get("is_admin"):
+            abort(401, message="You are not authorized to access this resource")
+
         item = ItemModel.query.get_or_404(item_id)
         db.session.delete(item)
         db.session.commit()
         return {"message": "Item deleted."}
 
-
+    # atualiza um item
+    @jwt_required()
     @blp.arguments(ItemUpdateSchema)
     @blp.response(200, ItemSchema)
     def put(self, item_data, item_id):
@@ -45,10 +54,14 @@ class Item(MethodView):
 
 @blp.route("/item")
 class ItemList(MethodView):
+    
+    # retorna todos os itens
     @blp.response(200, ItemSchema(many=True))
     def get(self):
         return ItemModel.query.all()
 
+    # cria um novo item
+    @jwt_required(fresh=True)
     @blp.arguments(ItemSchema)
     @blp.response(201, ItemSchema)
     def post(self, item_data):
